+++ src/typechecker/func.rs
@@ -151,22 +151,22 @@ pub fn check_func_def(
         let block = Block::check(checker, scope, &mut vars, func, ast_block)?;
 
         // check for return statement
-        if let Some(value) = &func.signature.returns {
-            if !block.value.iter().any(|stmnt| match stmnt {
-                Stmnt::Return(_) => true,
-                _ => false,
-            }) {
-                return Err(Error::MissingReturn {
-                    decl: func.span,
-                    name: value.name(checker),
-                    offending: Span::new(
-                        ast_block.span.file_id,
-                        ast_block.span.end - 1,
-                        ast_block.span.end,
-                    ),
-                });
-            }
-        }
+        // if let Some(value) = &func.signature.returns {
+        //     if !block.value.iter().any(|stmnt| match stmnt {
+        //         Stmnt::Return(_) => true,
+        //         _ => false,
+        //     }) {
+        //         return Err(Error::MissingReturn {
+        //             decl: func.span,
+        //             name: value.name(checker),
+        //             offending: Span::new(
+        //                 ast_block.span.file_id,
+        //                 ast_block.span.end - 1,
+        //                 ast_block.span.end,
+        //             ),
+        //         });
+        //     }
+        // }
 
         checker.funcs[item.0 as usize].func_impl = Some(FuncImpl { block, vars });
     }